src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.22.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.3
cryptography==39.0.2
distro==1.9.0
dnspython==2.3.0
eventlet==0.33.3
firebase-admin==6.7.0
Flask==2.2.2
Flask-SocketIO==5.3.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==2.0.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.8.2
MarkupSafe==2.1.2
msgpack==1.1.0
numpy==1.26.4
openai==1.60.0
opencv-python==4.9.0.80
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==23.0.0
pyparsing==3.2.3
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.32.3
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==2.2.2
